+++ lib/src/test/_index.dart
@@ -4,3 +4,4 @@
 
 export 'echo.dart';
 export 'json_data_service.dart';
+export 'save_as.dart';