+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Copyright (c) 2022, GM Consult Pty Ltd
 All rights reserved. 
 -->
 
+
+### 0.0.2
+
+#### New:
+- Added `SaveAs` class.
+- Added static method `SaveAs.text`.
+
 ### 0.0.2
 
 - Minor change to table formatting in `Echo` class